requirements.txt
amqp==2.5.1
asn1crypto==0.24.0
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.1.7
billiard==3.5.0.5
cached-property==1.3.1
celery==4.0.2
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
configparser==4.0.2
contextlib2==0.6.0.post1
coverage==4.5.1
cryptography==2.8
Django==1.10.4
djangorestframework==3.6.3
docker==2.5.1
docker-compose==1.17.1
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
Flask==1.1.1
funcsigs==1.0.2
functools32==3.2.3.post2
gyp==0.1
idna==2.5
importlib-metadata==0.23
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.5
MarkupSafe==1.1.1
mock==2.0.0
more-itertools==5.0.0
mysql-connector-python-rf==2.2.2
olefile==0.46
paramiko==2.6.0
pathlib2==2.3.5
pbr==3.1.1
Pillow==4.2.1
psutil==5.4.2
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pygobject==3.26.1
PyNaCl==1.3.0
pyOpenSSL==17.5.0
PyPDF2==1.26.0
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
redis==3.2.1
requests==2.18.1
scandir==1.10.0
SecretStorage==2.3.1
six==1.12.0
stormssh==0.7.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.21.1
vine==1.3.0
websocket-client==0.44.0
Werkzeug==0.16.0
zipp==0.6.0
